Fix puppet-lint namespace clash with new puppetlabs-puppet-lint gem
Currently, the new puppet-lint gem and the old one clash as both compete for the PuppetLint namespace. This initial commit aims to solve this by intorducing the puppetlabs directory in lib so that resources are found in lib/puppetlabs/puppet-lint rather than lib/puppet-lint.
This method is being tested first to avoid introducing changes that would require users to update their puppet-lint resource definitions