Avoid using the manifest lexer on YAML
A YAML file like in the fixture added is parsed fine by Psych and
yamllint, but puppet-lint chokes on it because it gets fed through the
manifest lexer first, which can't cope with some things that are valid
YAML.

Since the one check that operates on YAML files doesn't support fixing
and doesn't use tokens, it seems reasonable to bypass the lexer; more
sophisticated YAML checks would need to use a separate YAML lexer
anyway.

Note that this can't be a unit test because those bypass the lexer
processing.